% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneNames.R
\name{addGeneNames}
\alias{addGeneNames}
\title{Obtain gene names for each feature (currently, protein) in a given dataset in an spe object, add a column with gene names to the dataset and return a new spe object
Run this function before running any geneset based analysis functions if the starting data doesn't have a gene name column already.
Gene names are needed for any pathway/geneset based analysis as the members in each geneset are given as gene names, not protein identifiers.}
\usage{
addGeneNames(spe, featureID_colname, species, db_file)
}
\arguments{
\item{spe}{Spatial Experiment object containing the omics dataset}

\item{featureID_colname}{Column name containing feature names for which gene names are needed (Example: "Protein_rowID")}

\item{species}{Name of the species whose data is in spe (example: "human")}

\item{db_file}{Path or name of database file containing information for converting from feature ID to gene name. Defaults to database files provided in spammR for the given species}
}
\value{
spe_withGeneNames Spatial Experiment object containing everything in the input spe, along with a column for gene name.
}
\description{
Obtain gene names for each feature (currently, protein) in a given dataset in an spe object, add a column with gene names to the dataset and return a new spe object
Run this function before running any geneset based analysis functions if the starting data doesn't have a gene name column already.
Gene names are needed for any pathway/geneset based analysis as the members in each geneset are given as gene names, not protein identifiers.
}
