% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialProtUtils.R
\name{adjustPhophoWithGlobal}
\alias{adjustPhophoWithGlobal}
\title{adjustPhosphoWithGlobal
For some phospho analysis, we might want to investigate the phosphosites that are changing
indpendently of the proteomiocs data. Therefore we consume two independent data objects and subtract one from the other}
\usage{
adjustPhophoWithGlobal(phos.obj, prot.obj)
}
\arguments{
\item{phos.obj}{phosphoproteomic data}

\item{prot.obj}{proteomic data}
}
\value{
phos.obj
}
\description{
adjustPhosphoWithGlobal
For some phospho analysis, we might want to investigate the phosphosites that are changing
indpendently of the proteomiocs data. Therefore we consume two independent data objects and subtract one from the other
}
