% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff_pairwiseSamples.R
\name{calc_diff_pairwiseSamples}
\alias{calc_diff_pairwiseSamples}
\title{Helper function for computing pairwise differences between sample values
This function is called in distance_based_analysis()}
\usage{
calc_diff_pairwiseSamples(feature_values)
}
\arguments{
\item{feature_values}{A 1 x n vector containing values for a feature for n samples. names(feature_values) should contain names of samples}
}
\description{
Helper function for computing pairwise differences between sample values
This function is called in distance_based_analysis()
}
