% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_missingness.R
\name{calc_missingness}
\alias{calc_missingness}
\title{Calculate degree of missingness for each featurea and sample}
\usage{
calc_missingness(spe, group_name = NULL)
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{group_name}{Name of group to use as denominator}
}
\value{
SpatialExperiment object with missingFeature column added to colData and missingSample added to rowData
}
\description{
calc_missingness() calculates the fraction of samples that are missing data for each feature, and fraction of each feature that are missing for each sample.
}
\examples{
library(spammR)
data(smallPancData)
data(pancMeta)
data(protMeta)
pooledPanc <- dplyr::bind_cols(smallPancData)
panc.spe <- convert_to_spe(pooledPanc, pancMeta, protMeta, feature_meta_colname = "pancProts")
calc_missingness(panc.spe)

}
