% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSigScore.R
\name{calcSigScore}
\alias{calcSigScore}
\title{calcSigScore: calculates a signature score on a particular SingleCellExperiment object
This is a numeric score based on the mean rank of the proteins in teh signature}
\usage{
calcSigScore(sce, sigProts, sigName)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{sigProts}{list of gene names}

\item{sigNamme}{name of signature}
}
\description{
calcSigScore: calculates a signature score on a particular SingleCellExperiment object
This is a numeric score based on the mean rank of the proteins in teh signature
}
