% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_spe.R
\name{convert_to_spe}
\alias{convert_to_spe}
\title{Create a \code{SpatialExperiment} object from data}
\usage{
convert_to_spe(
  dat,
  sample_meta,
  feature_meta,
  assay_name = "proteomics",
  sample_colname = NULL,
  remove_samples = NULL,
  feature_data_colname = NULL,
  feature_meta_colname = NULL,
  spatialCoords_colnames = c("x_pixels", "y_pixels"),
  samples_common_identifier = "sample",
  image_files = NULL,
  image_ids = NULL,
  image_samples_common_identifier = NULL
)
}
\arguments{
\item{dat}{Matrix or data frame with omics measurements. Rows or \code{feature_colname} are features, columns are samples}

\item{sample_meta}{Data frame of metadata with either \code{sample_colname} or rownames as samples (if empty)}

\item{feature_meta}{Data frame of metadata with either \code{sample_colname} or rownames as samples (if empty)}

\item{assay_name}{Name to be given to the data in omics_measurements_file. Example: "abundance", "log2", "znormalized_log2" or any other descriptive name}

\item{sample_colname}{Column name in \code{sample_meta} table whose entries contain sample identifiers provided in \code{dat}.}

\item{remove_samples}{List of names of samples (as they occur in sample_colname) that should be removed/excluded from the data for SPE. Don't need to specify if no samples need to be removed.}

\item{feature_data_colname}{Name of column in \code{dat}, that is to be used for identifying features. If missing defaults to rownames.}

\item{feature_meta_colname}{Name of column in \code{feature_meta}, that is to be used for identifying features. If missing defaults to rownames.}

\item{spatialCoords_colnames}{A list containing names of columns in \code{meta_dat} that are spatial coordinates. Default value is NA which indicates that these columns are not present in the metadata_file, and in that case, the argument 'spatialCoords_file' must be specified.}

\item{samples_common_identifier}{A string (if same for all samples in omics_measurements_file) or a character vector (same length as number of samples in omics_measurements_file) corresponding to a descriptive name for samples in the current dataset. Examples: "Image0", "Experiment1", etc.}

\item{image_files}{A list containing paths of image files to be stored in the SpatialExperiment object. More images can be added later, without using this function.}

\item{image_ids}{A list containing image names/identifiers for image paths provided in image_files}

\item{image_samples_common_identifier}{A list containing names of samples_common_identifier(s) corresponding to images provided in image_files. This identifier links specific samples to a experiment/condition represented by a given image.}
}
\value{
spe.out a SpatialExperiment (SPE) object that contains all data and image(s). Ready to be used as input in spammR functions that require SPE object as input.
}
\description{
\code{convert_to_spe()} Puts omics data (omics measurements, metadata and image(s) corresponding to samples' tissue) into a SpatialExperiment (SPE) object. Most spammR functions require the input data to be an SPE object.
}
\examples{
\dontrun{
data(pancMeta)
data(protMeta)
data(pancDataList)
#We can put all samples into the same object (for statistical power)
pooledData<-dplyr::bind_cols(pancDataList)
pooled.panc.spe <- convert_to_spe(pooledData,pancMeta,protMeta,feature_meta_colname='pancProts',samples_common_identifier='')
#or we can add the inmage to a single data capture
img0.spe<-convert_to_spe(pancDataList$Image_0,pancMeta,protMeta,feature_meta_colname='pancProts',image_files=system.file("extdata",'Image_0.png',package='spammR'),image_samples_common_identifier='Image0',samples_common_identifier = 'Image0',image_ids='Image0')
}
}
