% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_based_analysis.R
\name{distance_based_analysis}
\alias{distance_based_analysis}
\title{Identify multiomic features correlated with an distance to an image feature}
\usage{
distance_based_analysis(
  spe,
  assayName,
  spotHeightCol = "spot_height",
  spotWidthCol = "spot_width",
  sampleCategoryCol,
  sampleCategoryValue,
  featuresNameCol,
  corr_type = "spearman",
  corr_thresh = 0.5,
  min_samples = 5,
  allowOverlaps = TRUE
)
}
\arguments{
\item{spe}{SpatialExperiment object containing spatial omics data}

\item{assayName}{Name of the assay stored in spe that is to be used for distance based analysis. Example: "znormalized_log2"}

\item{sampleCategoryCol}{Column name in metadata (colData(spe)) that should be used for selecting samples of certain type, to define the "origin" region for distance based analysis}

\item{sampleCategoryValue}{Sample category to be used for defining the "origin" region for distance based analysis}

\item{featuresNameCol}{Name of column containing features (example: proteins) in rowData(spe). It is assumed that the data provided in assay(spe) is in the same order as the order in which the features are listed under the featuresNameCol}

\item{corr_type}{Choose from "pearson" (default), "spearman." Correlation method to be used for calculating correlation between distance between samples and protein abundance differences. Both types of correlation provide a measure of monotonic association between two variables. Pearson is better suited for linear relationships while Spearman is better for non-linear monotonic relationships.}

\item{corr_thresh}{Minimum correlation value to be used for identifying proteins that have a correlation between protein abundance differences and distance between samples. Values greater than or equal to this theshold will be used.}

\item{sample_dimensions}{A vector containing the x and y dimensions of samples. Example: c(1,1) Sample Dimension units should match the units of spatial coordinates specified in spatialCoords(spe)}

\item{min_samplePoints_forCorr}{Requirement of a minimum number of sample points for calculating correlation. For proteins with less than this number of sample points, correlation value is reported as NA.}
}
\value{
dist_based_results List containing a.) pairwise_calculations_betweenSamples which is a list; each entry corresponds to a feature and contains a dataframe consisting of a distance (between samples) vector and a corresponding protein abundance difference vector.
b.) corr_pval_all and c.) corr_pval_thresholded which are data frames containing the correlation value, p-value and number of sample points used for the correlation calculation for each feature. corr_pval_thresholded only contains results with correlation > corr_thresh.
dist_based_results is also saved as .RData object under results_dir.
}
\description{
distance_based_analysis: Identifies proteins/features that show a strong correlation between distance from a specified ROI's samples and
protein/feature abundance differences between samples.
}
\examples{
data(pancData)
data(pancMeta)
data(pancDataList)
pooled.panc.spe <- convert_to_spe(pancData,pancMeta,protMeta,feature_meta_colname='pancProts',samples_common_identifier='')
img0.spe<-convert_to_spe(pancDataList$Image_0,pancMeta,protMeta,feature_meta_colname='pancProts',image_files=system.file("extdata",'Image_0.png',package='spammR'),image_samples_common_identifier='Image0',samples_common_identifier = 'Image0',image_ids='Image0')
img0.spe<-distance_based_analysis(img0.spe,'proteomics',sampleCategoryCol='IsletOrNot',sampleCategoryValue='Islet')
}
