% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_gradient.R
\name{enrich_gradient}
\alias{enrich_gradient}
\title{Calculate functional or pathway enrichment from a gradient}
\usage{
enrich_gradient(spe, geneset, feature_column, ranking_column)
}
\arguments{
\item{spe}{SpatialExperiment object containing spatial omics data and spatial diffex results}

\item{geneset}{in GMT format}

\item{feature_column}{Column of rowData that maps to gene set}

\item{ranking_column}{Column of rowData that maps to ranks}
}
\value{
A dataframe containing results from over-representation analysis of members of gene sets in the interest list of genes based on filtering criteria above.
}
\description{
\code{enrich_gradient()} calculates over-representation statistics (ORA) using a ranking of genes from \code{distance_based_analysis()}
For ORA using an external or already defined interest list of genes and gene sets, use leapR functions directly
}
\examples{
data(smallPancData)
data(pancMeta)
data(protMeta)
img0.spe <- convert_to_spe(smallPancData$Image_0,
            pancMeta,
            protMeta,
            feature_meta_colname = 'pancProts',
            image_files=system.file("extdata",'Image_0.png',package = 'spammR'),
            image_sample_ids = 'Image0',
            spatial_coords_colnames = c('x_pixels','y_pixels'),
            sample_id = 'Image0',
            image_ids='Image0')
img0.spe<-distance_based_analysis(img0.spe,
            'proteomics',
            sampleCategoryCol = 'IsletOrNot',
            sampleCategoryValue = 'Islet')
library(leapR)
data('krbpaths')
rank.res <- enrich_gradient(img0.spe, 
                geneset = krbpaths,
                feature_column = 'PrimaryGeneName',
                ranking_column = 'IsletDistancespearmanCor')
head(rank.res)
}
