% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_ORA.R
\name{enrich_ORA}
\alias{enrich_ORA}
\title{enrich_ORA: Does over-representation analysis (ORA) using an interest list of genes from differential expression results in spammR and gene sets (either the ones provided in spammR or user supplied)
This function uses results from spatialDiffEx.R and assumes input of a specific format. Interest list of genes for ORA is obtained from spatialDiffEx results based on the criteria
specified in this function.
For ORA using an external or already defined interest list of genes and gene sets, use leapR functions directly}
\usage{
enrich_ORA(
  spe,
  spatialDiffEx_results,
  pval_type_forThresh,
  pval_thresh,
  logFC_lowerThresh = NA,
  logFC_upperThresh = NA,
  geneset_spammR = "M2_curated_all",
  geneset_external_type = NA,
  geneset_external = NA,
  outdir,
  sortResultsBy,
  comparison_name
)
}
\arguments{
\item{spe}{SpatialExperiment object containing spatial omics data. This will be used to obtain background genes for ORA.}

\item{spatialDiffEx_results}{path to excel file containing results from spatialDiffEx.R}

\item{pval_type_forThresh}{Choose from "adjusted_pval" or "pval". Type of p-value that should be used for filtering statistically significant results. Default is adjusted p-value for multiple hypotheses correction.}

\item{pval_thresh}{value to use for filtering based on pval_type_forThreshold. Default is 0.05. Values less than pval_thresh will be kept.}

\item{logFC_lowerThresh}{Lower threshold for log Fold Change, to be used for filtering spatialDiffEx results. Default is NA}

\item{logFC_upperThresh}{Upper threshold for log Fold Change, to be used for filtering spatialDiffEx results. Default is NA}

\item{geneset_spammR}{Geneset to be used from spammR. Choose from: NA, "M2_curated_all" (default),"M2_cp_biocarta", "M2_cp_reactome","M5_GO_all", "M5_GO_bp", "M5_GO_cc", "M5_GO_mf"}

\item{geneset_external_type}{If not using a spammR provided geneset, specify type of geneset_external file. Options are ".gmt" or "list." If "list," make sure it is compatible with leapR format. Default is NA.}

\item{geneset_external}{If not using a spammR provided geneset, provide the filepath for the gene set file or the list object containing geneset information in leapR format. Default is NA.}

\item{sortResultsBy}{For sorting ORA results, choose from the following column names: "BH_pvalue" (default)}

\item{comparison_name}{Example: "RSPv_vs_others" Text to indicate in results data frame, which spatial groups were compared for the interest list of genes}

\item{ora_outdir}{Directory path for where results from ORA should be stored. Output is stored as an excel file as well as a dataframe as an Rdat file.}
}
\value{
a dataframe containing results from over-representation analysis of members of gene sets in the interest list of genes based on filtering criteria above.
}
\description{
enrich_ORA: Does over-representation analysis (ORA) using an interest list of genes from differential expression results in spammR and gene sets (either the ones provided in spammR or user supplied)
This function uses results from spatialDiffEx.R and assumes input of a specific format. Interest list of genes for ORA is obtained from spatialDiffEx results based on the criteria
specified in this function.
For ORA using an external or already defined interest list of genes and gene sets, use leapR functions directly
}
