% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expToLongForm.R
\name{expToLongForm}
\alias{expToLongForm}
\title{expToLongForm: helper function that moves expression
matrix to long form}
\usage{
expToLongForm(sce, rowname = "prot")
}
\description{
expToLongForm: helper function that moves expression
matrix to long form
}
