% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanUniprotIds}
\alias{humanUniprotIds}
\title{Table of human uniprot ids}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 207304 rows and 5 columns.
}
\usage{
humanUniprotIds
}
\description{
Table of human uniprot ids
}
\keyword{datasets}
