% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_data}
\alias{impute_data}
\title{spammR_impute: carries out imputation for missing data using a range of methods}
\usage{
impute_data(
  dat,
  method,
  knn_k = NULL,
  allowed_missingness_perProtein = NULL,
  allowed_missingness_perSample = NULL,
  metadata,
  spatial_unit_col
)
}
\arguments{
\item{dat}{Data frame containing data to be imputed}

\item{method}{Method of imputation to be used}

\item{metadata}{This should correspond to colData(spe) and should be specified if the imputation method is spatial specific}

\item{spatial_unit_col}{Column name that specifies spatial unit information. Example: ROI_abbreviation.}
}
\value{
a data frame containing the imputed dataset using the specified method.
}
\description{
spammR_impute: carries out imputation for missing data using a range of methods
}
\details{
Methods options and description
zero: replace missing values with 0
median_a : replace missing values with global median per protein
median_b : replace missing values with 1/2 global median per protein
global_mean: replace missing values with global mean per protein
mean_perSpatialUnit: replace msising values with mean per spatial unit (example: ROI) for each protein
knn_proteins_global: imputation based on k-nearest neighbors, with proteins as neighbors, based on data from all samples across all ROIs
knn_proteins_perSpatialUnit: imputation based on k-nearest neighbors, with proteins as neighbors, based on data from specified spatial unit (here: ROI)
knn_samples_global_proteinData: imputation based on k-nearest neighbors, with samples as neighbors, based on protein data from all samples.
knn_samples_global_spatialCoords: imputation based on knn, with samples as neighbors, based on spatial coordinates of all samples.
}
