% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_df}
\alias{impute_df}
\title{impute_df: carries out imputation for missing data in a data frame (df) using a specified method from a range of methods. Accepts input dataset as a dataframe (df).}
\usage{
impute_df(
  dat,
  method,
  metadata,
  spatial_unit_colname,
  spatialCoord_x_colname,
  spatialCoord_y_colname,
  knn_k = NULL,
  allowed_missingness_perProtein = NULL,
  allowed_missingness_perSample = NULL
)
}
\arguments{
\item{dat}{Data frame containing data to be imputed, where rows correspond to features (which can be specified as row names of dat but it is not required that they be specified) and columns correspond to samples. Column names must correspond to names provided in the sample identifier column in the metadata parameter.}

\item{method}{Method of imputation to be used. See details.}

\item{metadata}{A dataframe containing metadata for dat, where rows correspond to samples, and columns contain meta information about samples. This is required to be specified if the imputation method is spatial specific. At the very minimum, columns corresponding to sample identifier, spatial unit (if relevant) and spatial coordinates should be specified. Default is NULL.}

\item{spatial_unit_colname}{Column name in metadata that specifies spatial unit information. Example: ROI_abbreviation.}

\item{spatialCoord_x_colname}{Column name in metadata that specifies the X coordinate for each sample.}

\item{spatialCoord_y_colname}{Column name in metadata that specifies the Y coordinate for each sample.}
}
\value{
a data frame containing the imputed dataset using the specified method.
}
\description{
impute_df: carries out imputation for missing data in a data frame (df) using a specified method from a range of methods. Accepts input dataset as a dataframe (df).
}
\details{
Methods options and descriptions
zero: replace missing values with 0\n
median_a : replace missing values with global median per protein\n
median_b : replace missing values with 1/2 global median per protein\n
global_mean: replace missing values with global mean per protein\n
mean_perSpatialUnit: replace msising values with mean per spatial unit (example: ROI) for each protein\n
knn_proteins_global: imputation based on k-nearest neighbors, with proteins as neighbors, based on data from all samples across all ROIs\n
knn_proteins_perSpatialUnit: imputation based on k-nearest neighbors, with proteins as neighbors, based on data from specified spatial unit (here: ROI)\n
knn_samples_global_proteinData: imputation based on k-nearest neighbors, with samples as neighbors, based on protein data from all samples.\n
knn_samples_global_spatialCoords: imputation based on knn, with samples as neighbors, based on spatial coordinates of all samples.\n
}
