% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseUniprotIds}
\alias{mouseUniprotIds}
\title{Table of mouse uniprot ids}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 86411 rows and 7 columns.
}
\usage{
mouseUniprotIds
}
\description{
Table of mouse uniprot ids
}
\keyword{datasets}
