% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneKSoilsData}
\alias{oneKSoilsData}
\title{Soil samples from the 1000 soils project
Row names indiciate KEGG Orthologs, Column names indicate locations}
\format{
An object of class \code{data.frame} with 12438 rows and 110 columns.
}
\source{
\url{}
}
\usage{
oneKSoilsData
}
\description{
Soil samples from the 1000 soils project
Row names indiciate KEGG Orthologs, Column names indicate locations
}
\keyword{datasets}
