% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneKSoilsMeta}
\alias{oneKSoilsMeta}
\title{Soil samples from the 1000 soils project
Row names indiciate KEGG Orthologs, Column names indicate locations
Soil samples metadata for the sample information}
\format{
An object of class \code{data.frame} with 142 rows and 21 columns.
}
\usage{
data(oneKSoilsMeta)
}
\description{
Soil samples from the 1000 soils project
Row names indiciate KEGG Orthologs, Column names indicate locations
Soil samples metadata for the sample information
}
\keyword{datasets}
