% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancData}
\alias{pancData}
\title{Here is the data to be included in the spammR package}
\format{
An object of class \code{data.frame} with 6693 rows and 63 columns.
}
\usage{
pancData
}
\description{
Pancreatic proteomics measurements from Gosline et al. proteomics study.
Row names indicate individual proteins, columns indicate sample identifiers described in
\code{pancMeta}. Individual values show un-transformed abundance values for each protein.
}
\references{
\url{data_blah.com}
}
\keyword{datasets}
