% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancDataList}
\alias{pancDataList}
\title{List of proteomic measurements in a list sorted from individual regions of the pancreas.}
\format{
An object of class \code{list} of length 7.
}
\source{
\url{panc paper}
}
\usage{
pancDataList
}
\description{
List of proteomic measurements in a list sorted from individual regions of the pancreas.
}
\keyword{datasets}
