% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialProtUtils.R
\name{plotFeatureGrid}
\alias{plotFeatureGrid}
\title{plotFeatureGrid}
\usage{
plotFeatureGrid(sce, feats, featname)
}
\arguments{
\item{features}{are list of features to plo}
}
\description{
plots a numeric value of a single feature or set of features
}
