% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialProtUtils.R
\name{plotSigGrid}
\alias{plotSigGrid}
\title{plotSigGrid
plots a numeric value using the Xcoord and Ycoord columns
takes a numeric score from the colDAta}
\usage{
plotSigGrid(sce, sigName)
}
\arguments{
\item{sce}{SingleCellExperiment}
}
\description{
plotSigGrid
plots a numeric value using the Xcoord and Ycoord columns
takes a numeric score from the colDAta
}
