% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protMeta}
\alias{protMeta}
\title{Protein metadata for general uniprot data, including the pancProts
column for the panc data}
\format{
An object of class \code{data.frame} with 6662 rows and 6 columns.
}
\source{
Uniprot
}
\usage{
data(protMeta)
}
\description{
Protein metadata for general uniprot data, including the pancProts
column for the panc data
}
\keyword{datasets}
