% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metaspace_data.R
\name{retrieve_metaspace_data}
\alias{retrieve_metaspace_data}
\title{Retreive data objects from metspace}
\usage{
retrieve_metaspace_data(
  project_id = "2024-02-15_20h37m13s",
  fdr = 0.2,
  assay_name = "metabolites",
  sample_id = "sample"
)
}
\arguments{
\item{project_id}{Identifier for project}

\item{fdr}{FDR value above which to collect data}

\item{assay_name}{Name of assay to include in object}

\item{sample_id}{Name of sample}
}
\description{
\code{retrieve_metaspace_data()}Collects data from
http://metaspace2020.org by project id and returns it in
SpatialExperiment object
}
\examples{
mspe <- retrieve_metaspace_data()

name
}
