% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_reduce.R
\name{spat_reduce}
\alias{spat_reduce}
\title{Reduce a high resolution image measurement (origin) to a lower resultion
measurement (target)}
\usage{
spat_reduce(
  spe_target,
  spe_origin,
  origin_assay = "metabolomics",
  target_dims = c("spot_width", "spot_height"),
  origin_dims = c("spot_width", "spot_height")
)
}
\arguments{
\item{spe_target}{Low resolution image measurements}

\item{spe_origin}{High-resolution image measreuements to be mapped to target}

\item{target_dims}{Names of width and height of target spots, respectively}

\item{origin_dims}{Names of width and height of origin spots, respectively}

\item{stat}{Statistic used to summarize regions, choices are currently
'mean' and 'median'}
}
\value{
A copy of the spe_target object with the altExperiment slot containing
a reduced version of the origin
}
\description{
\code{spat_reduce()} Creates a single SpatialExperiment object
with all measurements mapped to the same coordinate space for plotting
and analysis
}
\details{
add detauls here
}
