% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_heatmap.R
\name{spatial_heatmap}
\alias{spatial_heatmap}
\title{spatial_heatmap: Creates a spatial heatmap for a given feature in a SpatialExperiment object (spe) and provides the option to label each sample in the x-y plane}
\usage{
spatial_heatmap(
  spe,
  feature,
  feature_type = "PrimaryGeneName",
  assay_name = "proteomics",
  plotBackground_img = TRUE,
  sample_id,
  image_id,
  image_boundaries,
  spatial_coord_type = "",
  spatial_coord_names = c("Xcoord", "Ycoord"),
  spot_size = c(1, 1),
  metric_display = "Protein abundance measure",
  label_column = NA,
  sample_label_color = "white",
  sample_label_size = 1.75,
  plot_title = NULL,
  interactive = FALSE
)
}
\arguments{
\item{spe}{SpatialExperiment (SPE) object}

\item{feature}{Name of the feature in the spe object whose values are to plotted in the spatial heat map. This should be a row name in rowData(spe)}

\item{feature_type}{Example: "GeneName". Default is whatever identifier is used in rownames. The name of feature_type must be present as a column in rowData(spe)}

\item{assay_name}{Name of assay in the spe object that contains data to be plotted}

\item{plotBackground_img}{Boolean (TRUE or FALSE) to indicate whether a background image should be plotted. Default is FALSE. If TRUE, the parameters image_boundaries and image_sample_id must be specified, and the image data must be present in the spe object, under imgData(spe)}

\item{sample_id}{The names of the background image's sample_id fields in the spe object; this provides a unique identifier for the background image to be used for plotting (if there are multiple images under imgData(spe)) and only plots samples associated with the specified image sample_id. Example: c("Image0","Raw_noMarkings"). Image data stored under the spe object can be viewed by imgData(spe)
.@param image_id The name of the background image image_id. Together with the \code{sample_id} this provides a unique imge to plot.}

\item{image_boundaries}{Background image's corners'coordinates. These are need to make sure that the background image lines up withe samples' coordinates correctly. Must specify in the following format: c(xmin_image, ymin_image, xmax_image, ymax_image). For example: c(0,0,21,25). These must be in the same coordinate system as the spatial coordinates for the samples in the SPE object (spatialCoords(spe)).}

\item{spatial_coord_type}{Position type for the given spatial coordinates of samples in spe. Current options are: "topleft_corner", "topright_corner". Default is blank, which assumes coordinates are bottom right.}

\item{spatial_coord_names}{Names of x and y spatial coordinates respectively in the spe. Example: c("Xcoord","Ycoord") or c("X","Y").}

\item{spot_size}{Is a vector of length 2 describing the width and height of each spot.}

\item{metric_display}{Legend title for spatial heatmap. If this parameter is not specified, legend title defaults to "Protein abundance measure"}

\item{label_column}{Column in colData(spe) to be used for labeling grid squares. If not specified, default is no labels.}

\item{sample_label_color}{Color to be used for labels of samples/grid squares. Default is white.}

\item{sample_label_size}{Font size for sample labels. Default is 1.75.}

\item{plot_title}{Title to be given to the spatial heatmap. Default is "Spatial signature for XYZ" where XYZ is the name of the specified feature}

\item{interactive}{Boolean value (TRUE/FALSE) indicating whether the plot should have interactive mouse hovering. If not specified, this defaults to TRUE. Note: grid squares can only be labeled when interactive = FALSE due to current ggplotly limitations.}
}
\value{
spatial_plot: Spatial heatmap of the chosen feature
}
\description{
spatial_heatmap: Creates a spatial heatmap for a given feature in a SpatialExperiment object (spe) and provides the option to label each sample in the x-y plane
}
\details{
Assumes that every sample (column) in assay(spe) has a corresponding x,y coordinate given in spatialCoords(spe)

Assumes that colData(spe) and spatialCoords(spe) are provided in the same order of samples (rows).

Assumes that the spe object contains background image data, stored in imgData(spe) if plotBackground_img is specified to be TRUE below.

Default values are specified for function parameters 'metric_display', 'label_column' and 'interactive,' if the user doesn't specify those.

Future versions of spatialPlot_feature() to include additional options for spatial_coord_type: "center", bottomleft_corner", "bottomright_corner"
}
\examples{
data(pancMeta)
data(pancDataList)
data(protMeta)
img0.spe<-convert_to_spe(pancDataList$Image_0,pancMeta,protMeta,feature_meta_colname='pancProts',image_files=system.file("extdata",'Image_0.png',package='spammR'),image_samples_common_identifier='Image0',spatialCoords_colnames=c('x_pixels','y_pixels'),samples_common_identifier = 'Image0',image_ids='with_grid')
res = spatial_heatmap(img0.spe, feature='INS', sample_id='Image0', image_id='with_grid', spatial_coord_names=c('x_pixels','y_pixels'), spot_size=unlist(colData(img0.spe)[1,c('spot_width','spot_height')]), image_boundaries=unlist(colData(img0.spe)[1,c('x_origin','y_origin','x_max','y_max')]),label_column='IsletOrNot', interactive=FALSE)
}
