% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDiffEx.R
\name{spatialDiffEx}
\alias{spatialDiffEx}
\title{spatialDiffEx: does differential expression using annotations in object}
\usage{
spatialDiffEx(sce, column = "pulpAnnotation", vals = c("red", "white"))
}
\description{
spatialDiffEx: does differential expression using annotations in object
}
