% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment.R
\docType{class}
\name{SpatialExperiment}
\alias{SpatialExperiment}
\alias{coerce,SpatialExperiment,SpatialExperiment-method}
\title{The spatial experiment class}
\usage{
SpatialExperiment(..., xCoord = c(), yCoord = c())
}
\value{
SpatialExperiment object
}
\description{
Based on the \linkS4class{RangedSummarizedExperiment} and \linksS4class{SingleCellExperiment} classes, this class
maintains omics measurements for various sptailly related regions alongside
their x and y-coordinates
}
