% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_spe.R
\name{split_spe}
\alias{split_spe}
\title{Divide SPE object into smaller objects by column features}
\usage{
split_spe(spe, split_colname, assay_name = NULL, feature_column = NULL)
}
\arguments{
\item{spe}{SpatialExperiment object containing spatial omics data and spatial diffex results}

\item{split_colname}{Column of rowData that maps to gene set}

\item{assay_name}{Optional name of assay to use in split data}
}
\value{
A list of SpatialExperiment objects containing a subset of the data
}
\description{
\code{split_spe()} splits an SPE object into smaller objects by a column feature. Ideally this shoudl be in the SpatialExperiment
class
}
\examples{
data(smallPancData)
data(pancMeta)
data(protMeta)
pooledPanc <- dplyr::bind_cols(smallPancData)
panc.spe <- convert_to_spe(pooledPanc,pancMeta,protMeta,feature_meta_colname='pancProts')
split_list <- split_spe(panc.spe,split_colname='Image')


}
