% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot_tab.R
\name{volcano_plot_tab}
\alias{volcano_plot_tab}
\title{Generate a volcano plot to visualize results from differential expression analysis stored in a data frame.
Use this function directly if differential expression results are in a separate data frame and not an spe object
Helper function for volcanoPlot_DiffExSpe.R; use volcanoPlot_DiffExSpe.R if differential expression results are in an spe object.
-log10(p-value) vs. log2(fold change)}
\usage{
volcano_plot_tab(
  diffEx_df,
  logFC_colname,
  pval_colname,
  pval_corrected,
  title,
  thresh = 0.05,
  sigLabel_colname
)
}
\arguments{
\item{diffEx_df}{A dataframe containing results from differential expression}

\item{logFC_colname}{column name in differenital expression results that represents log10 fold change}

\item{pval_colname}{column name in differential epxression results that represents the p-value to be plotted}

\item{pval_corrected}{Boolean indicating whether pval_colname represents the corrected p-value or not}

\item{title}{title for the plot}

\item{thresh}{Threshold for p-value to be used to annotate significant results on the plot}

\item{sigLabel_col}{Either a a vector with labels or a string (Example: "Gene") that is the column name in differential expression results that should be used for labeling significant results on the plot.}
}
\value{
ggplot Volcano plot
}
\description{
Generate a volcano plot to visualize results from differential expression analysis stored in a data frame.
Use this function directly if differential expression results are in a separate data frame and not an spe object
Helper function for volcanoPlot_DiffExSpe.R; use volcanoPlot_DiffExSpe.R if differential expression results are in an spe object.
-log10(p-value) vs. log2(fold change)
}
