% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_centers.R
\name{calc_centers}
\alias{calc_centers}
\title{Calculate centers of spots/samples for distance based analysis}
\usage{
calc_centers(spe)
}
\arguments{
\item{spe}{SpatialExperiment object containing omics data}
}
\description{
calc_centers calculates the centers of spots for numerical analysis
}
\examples{
if(FALSE) calc_centers()
}
