% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spatial_diff_ex.R
\name{calc_spatial_diff_ex}
\alias{calc_spatial_diff_ex}
\title{Identify differentially abundant features across image}
\usage{
calc_spatial_diff_ex(
  spe,
  assay_name = "proteomics",
  log_transformed = FALSE,
  category_col,
  compare_vals
)
}
\arguments{
\item{spe}{Spatial Experiment object containing data to be used for differential expression analysis}

\item{assay_name}{Name of the dataset stored in the spe object, that is to be used for the differential expression analysis. Example: znormalized_log2}

\item{log_transformed}{Is the data given in spe log2 transformed TRUE or FALSE}

\item{category_col}{Name of the column that specifies category of each sample. Example: "IsletOrNot"
#Categories from \code{category_col} will be compared in the differential expression analysis}

\item{compare_vals}{A vector containing names of categories from category_col to be compared. Only required if there are more than two values in \code{category_col}}
}
\value{
a list with diffEx.df, a data frame containing the differential expression results and diffEx.spe: Spatial Experiment object containing diffEx, stored in rowData(diffEx.spe)
}
\description{
\code{calc_spatial_diff_ex()} Calculates differential expression analysis using annotations in a SpatialExperiment object
}
\examples{
data(pancData)
data(pancMeta)
data(protMeta)
pooled.panc.spe <- convert_to_spe(pancData,pancMeta,protMeta,feature_meta_colname='pancProts',samples_common_identifier='')
diffex.spe <- calc_spatial_diff_ex(pooled.panc.spe,category_col='IsletOrNot')

}
