% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_spe.R
\name{convert_to_spe}
\alias{convert_to_spe}
\title{convert_to_spe: Puts omics data (omics measurements, metadata and image(s) corresponding to samples' tissue) into a SpatialExperiment (SPE) object. Most spammR function require the input data to be an SPE object.}
\usage{
convert_to_spe(
  omics_measurements_file,
  metadata_file,
  meta_colname_sampleIDs,
  remove_samples,
  feature_colname,
  spatilCoords_colnames,
  samples_common_identifier,
  image_files,
  image_ids,
  image_samples_common_identifier
)
}
\arguments{
\item{omics_measurements_file}{File path for omics measurements. Rows correspond to features. Samples are in columns. Can have additional columns also describing features.}

\item{metadata_file}{File path for metadata for samples. This can include spatial coordinates.}

\item{meta_colname_sampleIDs}{Column name in metadata_file whose entries contain sample identifiers provided in omics_measurements_file.}

\item{remove_samples}{Names of samples (as they occur in meta_colname_sampleIDs) that should be removed/excluded from the data for SPE.}

\item{feature_colname}{Name of column in omics_measurements_file, that is to be used for identifying features.}

\item{samples_common_identifier}{A string (if same for all samples in omics_measurements_file) or a character vector (same length as number of samples in omics_measurements_file) corresponding to a descriptive name for samples in the current dataset. Examples: "Image0", "Experiment1", etc.}

\item{image_files}{A list containing paths of image files to be stored in the SpatialExperiment object. More images can be added later, without using this function.}

\item{image_ids}{A list containing image names/identifiers for image paths provided in image_files}

\item{image_samples_common_identifier}{A list containing names of samples_common_identifier(s) corresponding to images provided in image_files. This identifier links specific samples to a experiment/condition represented by a given image.}

\item{assay_name}{Name to be given to the data in omics_measurements_file. Example: "abundance", "log2", "znormalized_log2" or any other descriptive name}

\item{spatialCoord_colnames}{A list containing names of columns in metadata_file that are spatial coordinates. Default value is NA which indicates that these columns are not present in the metadata_file.}

\item{spatialCoords_file}{If spatial coordinates are not provided in the metadata_file, then a file path for the spatial coordinates file should be specified.}
}
\value{
spe.out a SpatialExperiment (SPE) object that contains all data and image(s). Ready to be used as input in spammR functions that require SPE object as input.
}
\description{
convert_to_spe: Puts omics data (omics measurements, metadata and image(s) corresponding to samples' tissue) into a SpatialExperiment (SPE) object. Most spammR function require the input data to be an SPE object.
}
