% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_ora.R
\name{enrich_ora}
\alias{enrich_ora}
\title{enrich_ora: Does over-representation analysis (ORA) using an interest list of genes from differential expression results in spammR and gene sets (either the ones provided in spammR or user supplied)
This function uses results from spatialDiffEx.R and assumes input of a specific format. Interest list of genes for ORA is obtained from spatialDiffEx results based on the criteria
specified in this function.
For ORA using an external or already defined interest list of genes and gene sets, use leapR functions directly}
\usage{
enrich_ora(
  spe,
  geneset,
  feature_column,
  pval_type_forThresh = "adjusted_pval",
  pval_thresh = 0.05,
  logFC_lowerThresh = NA,
  logFC_upperThresh = NA,
  geneset_name = "msigdb",
  sortResultsBy,
  comparison_name = ""
)
}
\arguments{
\item{spe}{SpatialExperiment object containing spatial omics data and spatial diffex results}

\item{geneset}{in GMT format}

\item{feature_column}{Column of rowData that maps to gene set}

\item{pval_type_forThresh}{Choose from "adjusted_pval" or "pval". Type of p-value that should be used for
filtering statistically significant results. Default is adjusted p-value for multiple hypotheses correction.}

\item{pval_thresh}{value to use for filtering based on pval_type_forThreshold. Default is 0.05. Values less than pval_thresh will be kept.}

\item{logFC_lowerThresh}{Lower threshold for log Fold Change, to be used for filtering spatialDiffEx results. Default is NA}

\item{logFC_upperThresh}{Upper threshold for log Fold Change, to be used for filtering spatialDiffEx results. Default is NA}

\item{geneset_name}{Name of geneset provided}

\item{sortResultsBy}{For sorting ORA results, choose from the following column names: "BH_pvalue" (default)}

\item{comparison_name}{Example: "RSPv_vs_others" Text to indicate in results data frame, which spatial groups were compared for the interest list of genes}
}
\value{
a dataframe containing results from over-representation analysis of members of gene sets in the interest list of genes based on filtering criteria above.
}
\description{
enrich_ora: Does over-representation analysis (ORA) using an interest list of genes from differential expression results in spammR and gene sets (either the ones provided in spammR or user supplied)
This function uses results from spatialDiffEx.R and assumes input of a specific format. Interest list of genes for ORA is obtained from spatialDiffEx results based on the criteria
specified in this function.
For ORA using an external or already defined interest list of genes and gene sets, use leapR functions directly
}
\examples{

#data(pancData)
#data(pancMet)
#data(protMeta)
#panc.spe <- convert_to_spe(pancData,pancMeta,protMeta,
              feature_meta_colname='pancProts',samples_common_identifier='')
#diffex.spe <- calc_spatial_diff_ex(panc.spe,category_col='IsletOrNot',
               feature_colname='pancProts')
#library(leapR)
#data('msigdb')
#ora.spe <- enrich_ora(diffex.spe,geneset=msigdb,geneset_name='msigdb',
            feature_column='PrimaryGeneName')

}
