% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneKSoilsCoords}
\alias{oneKSoilsCoords}
\title{Soil sample coordinates for the sample information}
\format{
An object of class \code{data.frame} with 127 rows and 3 columns.
}
\usage{
oneKSoilsCoords
}
\description{
Soil sample coordinates for the sample information
}
\keyword{datasets}
