% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancMeta}
\alias{pancMeta}
\title{Sample metadata for the \code{pancData} file.}
\format{
An object of class \code{data.frame} with 63 rows and 7 columns.
}
\usage{
pancMeta
}
\description{
Sample metadata for the \code{pancData} file.
}
\keyword{datasets}
