% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancMeta}
\alias{pancMeta}
\title{Here is the data to be included in the spammR package}
\format{
An object of class \code{data.frame} with 63 rows and 15 columns.
}
\usage{
pancMeta
}
\description{
Sample metadata for the \code{pancData} file.
}
\references{
\url{panc paper}
}
\keyword{datasets}
