% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialProtUtils.R
\name{plotFeatureGrid}
\alias{plotFeatureGrid}
\title{plotFeatureGrid: plots a numeric value of a single feature or set of features}
\usage{
plotFeatureGrid(sce, feats, featname)
}
\description{
plotFeatureGrid: plots a numeric value of a single feature or set of features
}
