% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metaspace_data.R
\name{retrieve_metaspace_data}
\alias{retrieve_metaspace_data}
\title{Retreive data objects from metspace}
\usage{
retrieve_metaspace_data(
  project_id = "2024-02-15_20h37m13s",
  fdr = 0.2,
  assay_name = "lipids",
  sample_id = "sample",
  rotate = FALSE,
  drop_zeroes = TRUE,
  y_offset = 0,
  x_offset = 0
)
}
\arguments{
\item{project_id}{Identifier for project}

\item{fdr}{FDR value above which to collect data}

\item{assay_name}{Name of assay to include in object}

\item{sample_id}{Name of sample}

\item{rotate}{Set to TRUE if x and y coordinates need to be swapped}

\item{drop_zeroes}{Set to TRUE to drop zeroes and set the values to NAs. If
False the zeroes will be included in the image data. ß}

\item{y_offset}{Number of pixels to adjust y based on image}

\item{x_offset}{Number of pixels to adjust x based on image}
}
\description{
\code{retrieve_metaspace_data()}Collects data from
http://metaspace2020.org by project id and returns it in SpatialExperiment
object
}
\details{
This relies on underlying python code and environment to run.
}
\examples{
rdat <- retrieve_metaspace_data(project_id = "2024-02-15_20h37m13s", 
                                fdr = 0.2, 
                                assay_name = 'lipids', 
                               sample_id = 'sample')
head(rowMeans(assay(rdat,'lipids')))

}
