% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spammR-package.R
\docType{package}
\name{spammR-package}
\alias{spammR}
\alias{spammR-package}
\title{spammR: SPatial Analysis of Multiomics Measurements in R}
\description{
SpammR provides a suite of tools and a data structure to analyze diverse multiomics measurements in a spatial context. It enables the analysis and visualization of omics measurements paired with a specific image, each with spatial coordinates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pnnl-compbio.github.io/spammR/}
  \item Report bugs at \url{https://github.com/PNNL-CompBio/spammR/issues}
}

}
\author{
\strong{Maintainer}: Sara Gosline \email{sara.gosline@pnnl.gov} (\href{https://orcid.org/0000-0002-6534-4774}{ORCID})

Authors:
\itemize{
  \item Harkirat Sohi \email{harkirat.sohi@pnnl.gov} (\href{https://orcid.org/0009-0000-9935-1641}{ORCID})
  \item Yannick Mahlich \email{yannick.mahlich@pnnl.gov} (\href{https://orcid.org/0000-0002-0717-2927}{ORCID})
}

}
\keyword{internal}
