% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{spatial_network}
\alias{spatial_network}
\title{Plot network of correlated features}
\usage{
spatial_network(
  spe,
  assay_names,
  feature_names = NULL,
  features_of_interest = list(),
  method = "spearman"
)
}
\arguments{
\item{spe}{SpatialExperiment (SPE) object}

\item{assay_names}{Name of assay in the spe object that contains
data to be plotted. If only one it will focus on features in the first
assay. If more than one it will compute correlations across all features.}

\item{feature_names}{will describe which value to use for node names. IF
missing will use rownames}

\item{features_of_interest}{List of features to focus on. Should map to the
\code{feature_names} value. If empty will compute correlation across all features}

\item{method}{to use for correlation, default is \code{spearman}}
}
\value{
a \code{tidygraph} object that can be used for plotting or analysis
}
\description{
\code{spatial_network()} Creates a network for a given
SpatialExperiment object (spe) based on the correlation of features. Will
return a fully connect graph
}
\details{
Using the feature values, the algorithm computes the correlation
between them across space, filtes by \code{lowest_thresh} (in the interest of
compute time) and then selects the most correlated edges using the
\code{top_edges} parameter.

Only the largest connected component is plotted, but the entire
graph is returned as an \code{tidygraph} object.
}
