% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{spatial_network}
\alias{spatial_network}
\title{Plot network of correlated features}
\usage{
spatial_network(
  spe,
  assay_names,
  feature_names = NULL,
  features_of_interest = list(),
  lowest_thresh = 0.75,
  top_edges = 6000
)
}
\arguments{
\item{spe}{SpatialExperiment (SPE) object}

\item{assay_names}{Name of assay in the spe object that contains
data to be plotted. If only one it will focus on features in the first
assay. If more than one it will compute correlations across all features.}

\item{features_of_interest}{List of features to focus on. If empty will
compute correlation across
is used in rownames. The name of feature_type must be present as a
column in rowData(spe)}

\item{top_edges}{maximum number of correlation values to plot (for
better visibility)}

\item{feature}{Name of the feature in the spe object whose values are to
plotted in the spatial heat map. This should be a column name in
rowData(spe). If missing the rowNames will be used}
}
\description{
\code{spatial_network()} Creates a network for a given
SpatialExperiment object (spe) of the most correlated features within an
image
}
\details{
Using the feature values, the algorithm computes the correlation
between them across space, filtes by \code{lowest_thresh} (in the interest of
compute time) and then selects the most correlated edges using the
\code{top_edges} parameter.

Only the largest connected component is plotted, but the entire
graph is returned as an \code{tidygraph} object.
}
