% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment.R
\docType{class}
\name{SpatialExperiment-class}
\alias{SpatialExperiment-class}
\alias{SpatialExperiment}
\alias{coerce,SingleCellExperiment,SpatialExperiment-method}
\title{The spatial experiment class}
\usage{
SpatialExperiment(..., coords)
}
\value{
SpatialExperiment object
}
\description{
Based on the \linkS4class{RangedSummarizedExperiment} and \linksS4class{SingleCellExperiment} classes, this class
maintains omics measurements for various sptailly related regions alongside
their x and y-coordinates
}
\details{
Add details here
}
\examples{
data(pancData)
data(pancMeta)
spe<-SpatialExperiment(assays=list(logcounts=as(pancData),'dgCMatrix'),colData=pancMeta)
}
