% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot_feature.R
\name{spatialPlot_feature}
\alias{spatialPlot_feature}
\title{Creates a spatial heatmap for a given feature in a SpatialExperiment object (spe)
and provides the option to label each sample in the x-y plane}
\usage{
spatialPlot_feature(
  spe,
  spatial_coord_type,
  feature,
  metric_display = "Protein abundance measure",
  label_column = NA,
  interactive = TRUE
)
}
\arguments{
\item{spatial_coord_type}{Position type for the given spatial coordinates of samples in spe. Current options are: "topleft_corner", "topright_corner"
Future versions of spatialPlot_feature() to include additional options for spatial_coord_type: "center", bottomleft_corner", "bottomright_corner"}

\item{spe:}{SpatialExperiment object}

\item{feature:}{Name of the feature in the spe object whose values are to plotted in the spatial heat map. This should be a row name in rowData(spe)}

\item{metric_display:}{Legend title for spatial heatmap. If this parameter is not specified, legend title defaults to "Protein abundance measure"}

\item{label_column:}{Colunm in colData(spe) to be used for labeling grid squares. If not specified, default is no labels.}

\item{interactive:}{Boolean value (TRUE/FALSE) indicating whether the plot should have interactive mouse hovering. If not specified, this defaults to TRUE.}
}
\value{
spatial_plot: Spatial heatmap of the chosen feature
}
\description{
Creates a spatial heatmap for a given feature in a SpatialExperiment object (spe)
and provides the option to label each sample in the x-y plane
}
