% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Generate a volcano plot to visualize results from differential expression analysis.
In spammR, these could be results obtained from the spatialDiffEx() function.
log10(p-value) vs. log2(fold change)}
\usage{
volcano_plot(
  spe,
  logFC_col,
  pval_col,
  pval_corrected,
  title,
  thresh = 0.05,
  sigLabel_colname
)
}
\arguments{
\item{spe}{SpatialExperiment object containing results from differential expression}

\item{logFC_col}{column name in differenital expression results that represents log10 fold change}

\item{pval_col}{column name in differential epxression results that represents the p-value to be plotted}

\item{pval_corrected}{Boolean indicating whether the p-values to be used for thresholding significant results are corrected for multiple hypotheses or not.}

\item{title}{title for the plot}

\item{thresh}{Threshold for p-value to be used to annotate significant results on the plot}

\item{sigLabel_col}{Either a a vector with labels or a string (Example: "Gene") that is the column name in differential expression results that should be used for labeling significant results on the plot.}
}
\value{
ggplot Volcano plot
}
\description{
Generate a volcano plot to visualize results from differential expression analysis.
In spammR, these could be results obtained from the spatialDiffEx() function.
log10(p-value) vs. log2(fold change)
}
